test(graphql-transformers-e2e-tests): e2e test refactor + update
refactored mutation condition local e2e tests to proper file and updated e2e subscription e2e for
ownerfield as list
